pyproject.toml
[project]
name = "video-rag-project"
version = "0.1.0"
description = ""
authors = [
    {name = "derdentyler",email = "alexander.polybinsky@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "yt-dlp (>=2025.2.19,<2026.0.0)",
    "langdetect (>=1.0.9,<2.0.0)",
    "pysrt (>=1.1.2,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "youtube-transcript-api (>=1.0.1,<2.0.0)",
    "supabase (>=2.14.0,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "sentence-transformers (>=4.0.1,<5.0.0)",
    "llama-cpp-python @ https://github.com/abetlen/llama-cpp-python/releases/download/v0.2.90/llama_cpp_python-0.2.90-cp312-cp312-win_amd64.whl",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[tool.poetry]
packages = [{ include = "src" }]

[tool.poetry.scripts]
start-api = "src.api.server:main"
